+++ src/matrix/matrix.c
@@ -1,9 +1,9 @@
 #include "matrix.h"
 
-uint64_t **creat_new_matrix(int num_str, int str) {
+uint8_t **creat_new_matrix(int num_str, int str) {
   int p_trig=0;
-  uint64_t **mem_point = malloc(num_str * str * sizeof(uint64_t) + num_str * sizeof(uint64_t *));
-  uint64_t *p_mem = (uint64_t *)(mem_point + num_str);
+  uint8_t **mem_point = malloc(num_str * str * sizeof(uint8_t) + num_str * sizeof(uint8_t *));
+  uint8_t *p_mem = (uint8_t *)(mem_point + num_str);
   if (mem_point != NULL) {
     for (int i = 0; i < num_str; i++) {
       mem_point[i] = p_mem + str * i;
@@ -14,7 +14,7 @@ uint64_t **creat_new_matrix(int num_str, int str) {
   return (p_trig==0) ? mem_point : NULL;
 }
 
-uint64_t **fiil_matrix(uint64_t **matrix, char *message, int num_str, int str){
+uint8_t **fiil_matrix(uint8_t **matrix, char *message, int num_str, int str){
     for(int i = 0; i<num_str; i++){
         for(int j = 0; j<str; j++){
             matrix[i][j]=0;
@@ -31,6 +31,6 @@ uint64_t **fiil_matrix(uint64_t **matrix, char *message, int num_str, int str){
     return matrix;
 }
 
-void delete_matrix(uint64_t **matrix){
+void delete_matrix(uint8_t **matrix){
     free(matrix);
 }
